made performance changes

+++ types/fibonacci.go
@@ -2,6 +2,6 @@ package types
 
 type Fibonacci struct {
 	ID       int
-	Current  int
-	Previous int
+	Current  int64
+	Previous int64
 }
